run code cleanup

+++ src/main/java/com/github/burgerguy/hudtweaks/util/gl/GLUtil.java
@@ -14,7 +14,7 @@ import net.minecraft.util.math.Matrix4f;
 
 public enum GLUtil {
 	; // no instantiation, all contents static
-
+	
 	public static void drawBoxOutline(MatrixStack matrices, double x1, double y1, double x2, double y2, int color, float width) {
 		double j;
 		if (x1 < x2) {
@@ -22,17 +22,17 @@ public enum GLUtil {
 			x1 = x2;
 			x2 = j;
 		}
-		
+
 		if (y1 < y2) {
 			j = y1;
 			y1 = y2;
 			y2 = j;
 		}
-		
+
 		int a = color >> 24 & 255;
 		int r = color >> 16 & 255;
 		int g = color >> 8 & 255;
-		int b = color & 255;		
+		int b = color & 255;
 		Matrix4f matrix = matrices.peek().getModel();
 		VertexConsumer consumer = HTVertexConsumerProvider.getSolidOutlineConsumer(width);
 		consumer.vertex(matrix, (float) x1, (float) y2, 0.0F).color(r, g, b, a).next();
@@ -41,7 +41,7 @@ public enum GLUtil {
 		consumer.vertex(matrix, (float) x1, (float) y1, 0.0F).color(r, g, b, a).next();
 		HTVertexConsumerProvider.draw();
 	}
-	
+
 	public static void drawFillColor(MatrixStack matrices, double x1, double y1, double x2, double y2, int color) {
 		double j;
 		if (x1 < x2) {
@@ -49,17 +49,17 @@ public enum GLUtil {
 			x1 = x2;
 			x2 = j;
 		}
-		
+
 		if (y1 < y2) {
 			j = y1;
 			y1 = y2;
 			y2 = j;
 		}
-		
-		float f = (float) (color >> 24 & 255) / 255.0F;
-		float g = (float) (color >> 16 & 255) / 255.0F;
-		float h = (float) (color >> 8 & 255) / 255.0F;
-		float k = (float) (color & 255) / 255.0F;
+
+		float f = (color >> 24 & 255) / 255.0F;
+		float g = (color >> 16 & 255) / 255.0F;
+		float h = (color >> 8 & 255) / 255.0F;
+		float k = (color & 255) / 255.0F;
 		BufferBuilder bufferBuilder = Tessellator.getInstance().getBuffer();
 		RenderSystem.enableBlend();
 		RenderSystem.disableTexture();
